Split entity system further
